+++ psutil/_psutil_bsd.c
@@ -827,7 +827,11 @@ psutil_users(PyObject *self, PyObject *args) {
             py_tty,             // tty
             py_hostname,        // hostname
             (float)ut.ut_time,  // start time
+#ifdef PSUTIL_OPENBSD
+            -1                  // process id (set to None later)
+#else
             ut.ut_pid           // process id
+#endif
         );
         if (!py_tuple) {
             fclose(fp);
@@ -865,7 +869,11 @@ psutil_users(PyObject *self, PyObject *args) {
             py_tty,        // tty
             py_hostname,   // hostname
             (float)utx->ut_tv.tv_sec,  // start time
-            utx->ut_pid  // process id
+#ifdef PSUTIL_OPENBSD
+            -1             // process id (set to None later)
+#else
+            utx->ut_pid    // process id
+#endif
         );
 
         if (!py_tuple) {