+++ psutil/arch/bsd/openbsd.c
@@ -479,15 +479,21 @@ psutil_inet6_addrstr(struct in6_addr *p)
 }
 
 
+/*
+ * List process connections.
+ * Note: there is no net_connections() on OpenBSD. The Python
+ * implementation will iterate over all processes and use this
+ * function.
+ * Note: local and remote paths cannot be determined for UNIX sockets.
+ */
 PyObject *
 psutil_proc_connections(PyObject *self, PyObject *args) {
     long pid;
-    int i, cnt;
-
+    int i;
+    int cnt;
     struct kinfo_file *freep = NULL;
     struct kinfo_file *kif;
     char *tcplist = NULL;
-
     PyObject *py_retlist = PyList_New(0);
     PyObject *py_tuple = NULL;
     PyObject *py_laddr = NULL;
@@ -608,15 +614,18 @@ psutil_proc_connections(PyObject *self, PyObject *args) {
                     goto error;
                 Py_DECREF(py_tuple);
             }
-            // UNIX socket
+            // UNIX socket.
+            // XXX: local addr is supposed to be in "unp_path" but it
+            // always empty; also "fstat" command is not able to show
+            // UNIX socket paths.
             else if (kif->so_family == AF_UNIX) {
                 py_tuple = Py_BuildValue(
-                    "(iiisOi)",
+                    "(iiissi)",
                     kif->fd_fd,
                     kif->so_family,
                     kif->so_type,
-                    kif->unp_path,
-                    Py_None,
+                    "",  // laddr (kif->unp_path is empty)
+                    "",  // raddr
                     PSUTIL_CONN_NONE);
                 if (!py_tuple)
                     goto error;