+++ psutil/tests/test_misc.py
@@ -22,7 +22,6 @@ import stat
 import sys
 
 from psutil import LINUX
-from psutil import OSX
 from psutil import POSIX
 from psutil import WINDOWS
 from psutil._common import memoize
@@ -30,14 +29,21 @@ from psutil._common import memoize_when_activated
 from psutil._common import supports_ipv6
 from psutil._compat import PY3
 from psutil.tests import APPVEYOR
+from psutil.tests import bind_socket
 from psutil.tests import bind_unix_socket
+from psutil.tests import call_until
 from psutil.tests import chdir
 from psutil.tests import create_proc_children_pair
 from psutil.tests import create_sockets
 from psutil.tests import get_free_port
 from psutil.tests import get_test_subprocess
+from psutil.tests import HAS_MEMORY_FULL_INFO
 from psutil.tests import HAS_MEMORY_MAPS
+from psutil.tests import HAS_SENSORS_BATTERY
+from psutil.tests import HAS_SENSORS_FANS
+from psutil.tests import HAS_SENSORS_TEMPERATURES
 from psutil.tests import importlib
+from psutil.tests import is_namedtuple
 from psutil.tests import mock
 from psutil.tests import reap_children
 from psutil.tests import retry
@@ -459,7 +465,7 @@ class TestScripts(unittest.TestCase):
     def test_pmap(self):
         self.assert_stdout('pmap.py', args=str(os.getpid()))
 
-    @unittest.skipIf(not OSX or WINDOWS or LINUX, "platform not supported")
+    @unittest.skipIf(not HAS_MEMORY_FULL_INFO, "not supported")
     def test_procsmem(self):
         self.assert_stdout('procsmem.py')
 
@@ -486,30 +492,20 @@ class TestScripts(unittest.TestCase):
     def test_cpu_distribution(self):
         self.assert_syntax('cpu_distribution.py')
 
+    @unittest.skipIf(not HAS_SENSORS_TEMPERATURES, "not supported")
     @unittest.skipIf(TRAVIS, "unreliable on TRAVIS")
     def test_temperatures(self):
-        if hasattr(psutil, "sensors_temperatures") and \
-                psutil.sensors_temperatures():
-            self.assert_stdout('temperatures.py')
-        else:
-            self.assert_syntax('temperatures.py')
+        self.assert_stdout('temperatures.py')
 
+    @unittest.skipIf(not HAS_SENSORS_FANS, "not supported")
     @unittest.skipIf(TRAVIS, "unreliable on TRAVIS")
     def test_fans(self):
-        if hasattr(psutil, "sensors_fans") and psutil.sensors_fans():
-            self.assert_stdout('fans.py')
-        else:
-            self.assert_syntax('fans.py')
+        self.assert_stdout('fans.py')
 
+    @unittest.skipIf(not HAS_SENSORS_BATTERY, "not supported")
     def test_battery(self):
-        if hasattr(psutil, "sensors_battery") and \
-                psutil.sensors_battery() is not None:
-            self.assert_stdout('battery.py')
-        else:
-            self.assert_syntax('battery.py')
+        self.assert_stdout('battery.py')
 
-    @unittest.skipIf(APPVEYOR or TRAVIS, "unreliable on CI")
-    @unittest.skipIf(OSX, "platform not supported")
     def test_sensors(self):
         self.assert_stdout('sensors.py')
 
@@ -619,6 +615,10 @@ class TestSyncTestUtils(unittest.TestCase):
         wait_for_file(TESTFN, delete=False)
         assert os.path.exists(TESTFN)
 
+    def test_call_until(self):
+        ret = call_until(lambda: 1, "ret == 1")
+        self.assertEqual(ret, 1)
+
 
 class TestFSTestUtils(unittest.TestCase):
 
@@ -686,6 +686,11 @@ class TestProcessUtils(unittest.TestCase):
 
 class TestNetUtils(unittest.TestCase):
 
+    def bind_socket(self):
+        port = get_free_port()
+        with contextlib.closing(bind_socket(addr=('', port))) as s:
+            self.assertEqual(s.getsockname()[1], port)
+
     @unittest.skipIf(not POSIX, "POSIX only")
     def test_bind_unix_socket(self):
         with unix_socket_path() as name:
@@ -745,6 +750,13 @@ class TestNetUtils(unittest.TestCase):
                 self.assertGreaterEqual(fams[socket.AF_UNIX], 2)
             self.assertGreaterEqual(types[socket.SOCK_STREAM], 2)
             self.assertGreaterEqual(types[socket.SOCK_DGRAM], 2)
+
+
+class TestOtherUtils(unittest.TestCase):
+
+    def test_is_namedtuple(self):
+        assert is_namedtuple(collections.namedtuple('foo', 'a b c')(1, 2, 3))
+        assert not is_namedtuple(tuple())
 
 
 if __name__ == '__main__':